Rename view file
