+++ src/api/views/__init__.py
@@ -1,6 +1,6 @@
 from .auth import PublicKeyView
 from .credentials import CredentialsView
-from .subscriptions import OutgoingSubscriptionList
+from .subscription import OutgoingSubscriptionList
 from .user import UserDetail, UserList, UserSearch
 
 __all__ = [
